+++ src/renderer/Common/Icons/GnomeMaximize.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  let { size = "24" } = $props();
+  let { size = "24", color = "currentColor" } = $props();
 </script>
 
 <svg
@@ -10,6 +10,6 @@
   xmlns="http://www.w3.org/2000/svg"
 >
   <g transform="translate(4, 4)">
-    <path d="M3.98651 3.99219V4.99219V12.0039H11.9982V3.99219H3.98651ZM5.98651 5.99219H9.99823V10.0039H5.98651V5.99219Z" fill="#E6E6E7"/>
+    <path d="M3.98651 3.99219V4.99219V12.0039H11.9982V3.99219H3.98651ZM5.98651 5.99219H9.99823V10.0039H5.98651V5.99219Z" fill={color}/>
   </g>
 </svg>